+++ swap/httpswap/metrics.go
@@ -1,4 +1,4 @@
-package httpnet
+package httpswap
 
 import (
 	"context"
