+++ swap/httpswap/msg_sender.go
@@ -1,4 +1,4 @@
-package httpnet
+package httpswap
 
 import (
 	"bytes"
